requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
caldav==1.3.9
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
python-dotenv==1.0.0
